requirements.txt
flask
flask-cors
tensorflow
keras
ultralytics
pillow
numpy 